requirements.txt
# Core Dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=0.19.0

# Database & Caching
sqlalchemy>=1.4.23
aiohttp>=3.8.0
aiosqlite>=0.17.0
redis>=4.1.0

# Error Recovery & Resilience
tenacity>=8.0.1
circuitbreaker>=1.4.0

# Monitoring & Observability
prometheus-client>=0.13.1
opentelemetry-api>=1.9.1
opentelemetry-sdk>=1.9.1
opentelemetry-instrumentation-fastapi>=0.31b0
opentelemetry-exporter-prometheus~=0.55b1

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
cryptography>=36.0.0

# Utilities
pydantic>=1.8.2
python-dateutil>=2.8.2
structlog>=21.1.0
orjson>=3.6.4

# Testing
pytest>=6.2.5
pytest-asyncio>=0.15.1
httpx>=0.22.0
pytest-cov>=2.12.1

# Development
black>=21.12b0
isort>=5.10.1
mypy>=0.931
flake8>=4.0.1

# AI/ML (add as needed)
torch>=1.10.0
transformers>=4.12.0
numpy>=1.21.0
pandas>=1.3.0
